Indexer automatic versioning tuning
It might be needed to adjust references pointing to the old resource
version so that they point to the new one. Still, this requires the new
version resource to already exist, therefore can not be done by the
handler preparing to-be-created new version metadata. This requires a
dedicated handler. This commit introduces such a handler to the
Indexer::setVersioning(), File::upload() and File::uploadAsync(). The
handler's signature is `function(acdhOeaw\arche\lib\RepoResource $old,
acdhOeaw\arche\lib\RepoResource $new): void`